+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `list_active_tokens` Ecto admin function
     - `Boruta.AccessTokensAdapter`, `Boruta.CodesAdapter`, `Boruta.ClientsAdapter`, and `Boruta.ScopesAdapter` encapsulating adapters that are set in configuration.
     - `Boruta.Oauth.AuthorizeResponse.redirect_to_url/1` function
+    - `Boruta.Oauth.Error.redirect_to_url/1` function
     - `boruta.gen.migration` mix task
     - `Boruta.Ecto` schemas documentation
   ### Security